State and state manager packages
The state and state manager packages adds vertical boundaries to a given
state access. The idea behind this is to consolidate various
commonalities between various pieces of state into one location, so that
it becomes a lot easier to manage.

The first state manager package is the schema state, this handles the
upgrading of a state package.